Adicionando cpf no formulário de register e adicionando cpf na tabela do banco de dados, alterando value de float para double, iniciando tratamentos com cpf e mudando index de cada carro na lista

+++ src/screens/AddCarScreen.java
@@ -105,7 +105,7 @@ public class AddCarScreen extends JFrame{
         String brand = brandInput.getText();
         String model = modelInput.getText();
         int manufactureYear = Integer.parseInt(manufactureYearInput.getText());
-        float value = Float.parseFloat(valueInput.getText());
+        Double value = Double.parseDouble(valueInput.getText());
 
         
         boolean result = d.addCar(idUser, brand, model, manufactureYear, value);
